refactor(client): split monolithic types.ts into domain modules
Split 1541-line lib/types.ts into 18 domain files under lib/types/:
common, user, guild, channel, message, voice, auth, admin, pages,
social, e2ee, events, preferences, search, settings, favorites,
observability. index.ts re-exports all so existing
`@/lib/types` imports continue to work.
